reshuffle formatting
