+++ packages/studio-server/src/preview-server/routes/log-updates/format-side-props.ts
@@ -1,0 +1,78 @@
+import {RenderInternals} from '@remotion/renderer';
+import {
+	attrName,
+	colorValue,
+	equals,
+	punctuation,
+	strikeThrough,
+} from './formatting';
+
+export type PropDelta = {
+	key: string;
+	valueString: string;
+};
+
+const colorEnabled = () => RenderInternals.chalk.enabled();
+
+// Format key={value} with Monokai syntax highlighting
+const formatSimpleProp = (key: string, value: string) => {
+	return `${attrName(key)}${equals('=')}${punctuation('{')}${colorValue(value)}${punctuation('}')}`;
+};
+
+// Format parentKey={{childKey: value}} with Monokai syntax highlighting
+const formatNestedProp = (
+	parentKey: string,
+	childKey: string,
+	value: string,
+) => {
+	return `${attrName(parentKey)}${equals('=')}${punctuation('{{')}${punctuation(childKey)}${punctuation(':')} ${colorValue(value)}${punctuation('}}')}`;
+};
+
+const formatPropDelta = ({key, valueString}: PropDelta) => {
+	if (!colorEnabled()) {
+		const dotIndex = key.indexOf('.');
+		if (dotIndex === -1) {
+			return `${key}={${valueString}}`;
+		}
+
+		const parent = key.slice(0, dotIndex);
+		const child = key.slice(dotIndex + 1);
+		return `${parent}={{${child}: ${valueString}}}`;
+	}
+
+	const dotIdx = key.indexOf('.');
+	if (dotIdx === -1) {
+		return formatSimpleProp(key, valueString);
+	}
+
+	return formatNestedProp(
+		key.slice(0, dotIdx),
+		key.slice(dotIdx + 1),
+		valueString,
+	);
+};
+
+export const formatSideProps = ({
+	removedProps,
+	addedProps,
+}: {
+	removedProps: PropDelta[];
+	addedProps: PropDelta[];
+}) => {
+	const parts: string[] = [];
+
+	for (const prop of removedProps) {
+		const formatted = formatPropDelta(prop);
+		parts.push(colorEnabled() ? strikeThrough(formatted) : formatted);
+	}
+
+	for (const prop of addedProps) {
+		parts.push(formatPropDelta(prop));
+	}
+
+	if (parts.length === 0) {
+		return '';
+	}
+
+	return `, ${parts.join(', ')}`;
+};
